Added a Utils class
